# Optimize GPT-2 regex as logic for improved performance