--- a/tokenizers/benches/bpe_benchmark.rs
+++ b/tokenizers/benches/bpe_benchmark.rs
@@ -10,13 +10,16 @@ use std::path::Path;
 use criterion::Criterion;
 use tokenizers::models::bpe::{BpeTrainerBuilder, BPE};
 use tokenizers::models::TrainerWrapper;
-use tokenizers::pre_tokenizers::byte_level::ByteLevel;
+use tokenizers::pre_tokenizers::byte_level::{gpt2_regex_optimized, ByteLevel};
 use tokenizers::pre_tokenizers::whitespace::Whitespace;
 use tokenizers::tokenizer::{AddedToken, EncodeInput};
-use tokenizers::Tokenizer;
+use tokenizers::{NormalizedString, SplitDelimiterBehavior, Tokenizer};
 
 use common::{iter_bench_encode, iter_bench_encode_batch, iter_bench_train};
+use lazy_static::lazy_static;
+use onig::Regex;
 use std::ops::Deref;
+use std::time::Duration;
 
 static BATCH_SIZE: usize = 1_000;
 
@@ -100,6 +103,45 @@ fn bench_train(c: &mut Criterion) {
     });
 }
 
+lazy_static! {
+    static ref RE: Regex =
+        Regex::new(r"'s|'t|'re|'ve|'m|'ll|'d| ?\p{L}+| ?\p{N}+| ?[^\s\p{L}\p{N}]+|\s+(?!\S)|\s+")
+            .unwrap();
+}
+
+fn bench_gpt_regex(c: &mut Criterion) {
+    let re_ref: &Regex = &RE;
+    let mut lines: Vec<NormalizedString> = vec![];
+    for line in BufReader::new(File::open(Path::new("data/big.txt")).unwrap()).lines() {
+        let input = line.unwrap();
+        lines.push(NormalizedString::from(input.trim().clone()));
+    }
+    c.bench_function("BPE split REGEX", |b| {
+        b.iter(|| {
+            for normalized in &lines {
+                normalized
+                    .split(re_ref, SplitDelimiterBehavior::Isolated)
+                    .unwrap();
+            }
+        })
+    });
+}
+
+fn bench_gpt_logic(c: &mut Criterion) {
+    let mut lines: Vec<NormalizedString> = vec![];
+    for line in BufReader::new(File::open(Path::new("data/big.txt")).unwrap()).lines() {
+        let input = line.unwrap();
+        lines.push(NormalizedString::from(input.trim().clone()));
+    }
+    c.bench_function("BPE split Logic", |b| {
+        b.iter(|| {
+            for normalized in &lines {
+                gpt2_regex_optimized(&normalized).unwrap();
+            }
+        })
+    });
+}
+
 criterion_group! {
     name = benches;
     config = Criterion::default().sample_size(20);
@@ -110,4 +152,14 @@ criterion_group! {
     config = Criterion::default().sample_size(10);
     targets = bench_train
 }
-criterion_main!(benches, benches_train);
+criterion_group! {
+    name = benches_regex;
+    config = Criterion::default().sample_size(100).warm_up_time(Duration::from_secs(10));
+    targets = bench_gpt_regex
+}
+criterion_group! {
+    name = benches_regex_logic;
+    config = Criterion::default().sample_size(100).warm_up_time(Duration::from_secs(10));
+    targets = bench_gpt_logic
+}
+criterion_main!(benches, benches_train, benches_regex, benches_regex_logic);