The regex used by the BPE pre-tokenizer has fairly simple logic. The tokens are split along: letters, numbers, whitespace, symbols. Each possibly being prefixed with a `' '` and with special handling for specific apostrophe scenarios. 

Transforming this into logic instead of REGEX improves performance. Here are benchmarks (for the split function only):

Ran with 10s warmup over 100 samples of the `big.txt` data set.
```
BPE split REGEX         time:   [1.1157 s 1.1181 s 1.1207 s]

BPE split Logic         time:   [640.42 ms 645.15 ms 650.22 ms]
```

Here are some benchmarks for the overall GPT2 process (I include only the improvements, the previous run was the REGEX split)

```
BPE GPT2 encode         time:   [29.143 us 29.432 us 29.861 us]
                        change: [-12.354% -10.057% -7.6869%] (p = 0.00 < 0.05)
                        Performance has improved.

BPE GPT2 encode batch   time:   [8.8120 ms 9.0704 ms 9.3106 ms]
                        change: [-7.4765% -5.5234% -3.6632%] (p = 0.00 < 0.05)
                        Performance has improved.

BPE GPT2 encode, no cache
                        time:   [35.395 us 35.681 us 36.070 us]
                        change: [-14.634% -13.082% -11.349%] (p = 0.00 < 0.05)
                        Performance has improved.

BPE GPT2 encode batch, no cache
                        time:   [10.708 ms 10.824 ms 10.952 ms]
                        change: [-10.197% -7.7006% -5.4115%] (p = 0.00 < 0.05)
                        Performance has improved.
```

All the benchmarks were ran on my local machine:

```
MacBook Pro (16-inch, 2019)
2.4 GHz 8-Core Intel Core i9
64 GB 2667 MHz DDR4
Intel UHD Graphics 630 1536 MB
```