@Narsil 2022-04-06
Hi @benwtrent ,

Thanks very much for this PR !
Do you mind sharing the initial problem you are trying to solve ?

`10%` speedup is great ! But I fear the two codes don't necessarily do the same thing and it would need a lot more tests on the unicode range + making sure the codes do exactly the same thing.
Also speedups need to be checked on various platforms (even if we limit it to the most common ones)

Before investing tons of work to make sure things are 100% similar, can we first make the case for such a change (currently `tokenizers` should be a tiny amount of compute compared to the models that run with them) ?

Also another option would be to disable the regex `ByteLevel(use_regex=False)` and use a new `pre_tokenization` layer in `Sequence` so we can do the change more seemlessly.