@benwtrent 2022-04-06
> Do you mind sharing the initial problem you are trying to solve ?

Speed improvements. Even onig::Regex suffers when text input is large. And this saves compute. Note: the 10% is only on the `big.txt`, of which the longest input string is only 465 words. I postulate that for longer inputs this improvement is even greater. I can try to gather some numbers around improvements on longer text inputs.

> Before investing tons of work to make sure things are 100% similar, can we first make the case for such a change 

Again, long text inputs are a big pain for BPE, specifically because the regex match can be slow, thus hampering throughput.

> Also another option would be to disable the regex

Would this contribution be more readily accepted if I added a new pre-tokenization layer that is optionally available? Maybe one that users could optionally use if they know their input texts are usually exceptionally long?

I 100% understand the concern that the tokens created are not exactly the same in all cases. I can happily add more test cases.