@Narsil 2022-04-06
> Would this contribution be more readily accepted if I added a new pre-tokenization layer that is optionally available? Maybe one that users could optionally use if they know their input texts are usually exceptionally long?

That would make it much easier to add yes, since it would be only an opt-in feature.
We would still need to make sure both are the same, but it would definitely be easier to do.

Just to emphasize:

```python
tokenizer.pre_tokenizer = ByteLevel()

# Into
tokenizer.pre_tokenizer = Sequence([OptimizedByteRegex(), ByteLevel(use_regex=False)])
```

I am still confused how `onig` or other optimized regex engine can be "slow" on long text, afaik, there are also using a DFA just like the one you crafted. I know Python regex engine is not great at optimizing its DFA, but `onig` should do a better job at it, so I am not sure to understand why this is faster.


> I 100% understand the concern that the tokens created are not exactly the same in all cases. I can happily add more test cases.

The cases that pop into mind would be odd utf-8 whitespace, I sort of remember some categories could overlap  but cannot find relevant doc atm. Also making sure `onig` and `std::char` respect the same standard. Rust declares it follows https://docs.rs/regex/latest/regex/#unicode UTS#18 I am unsure about `onig`.