@benwtrent 2022-04-06
@Narsil 

> I am still confused how onig or other optimized regex engine can be "slow" on long text, afaik, there are also using a DFA just like the one you crafted. 

OK, I did some more tests. Specifically, I concatenated every 1000 lines from `big.txt` to dramatically increase the length of every example. 

The percentage improvements on the "optimized" vs regular onig REGEX were the same as before. onig is very clever and efficient. 

There goes my hypothesis 🤦 .

Well, this is great news for tokenizers :D. No need for this PR. 


FYI, working in this code base is a joy. Super easy to set up, run formatting, compiling, etc. ++ developer experience.