@benwtrent 2022-04-06
@Narsil the only "big" improvement I can think of that this logic provides is the ability to lazily stream tokens before collecting them. The REGEX has to read the entire string, where the logic can just read a handful of chars at a time and produce tokens as they are read.

However, it seems that this is not a goal of tokenizers and would be much more complicated to implement as other code would have to be changed (traits updated, etc.).