@Narsil 2022-04-07
> FYI, working in this code base is a joy. Super easy to set up, run formatting, compiling, etc. ++ developer experience.

Music to our ears !
🤗 

> @Narsil the only "big" improvement I can think of that this logic provides is the ability to lazily stream tokens before collecting them. The REGEX has to read the entire string, where the logic can just read a handful of chars at a time and produce tokens as they are read.

This is a fantastic idea, but probably hard to implement for sure, currently the tokenizer does have to do n passes on the strings to finish it's calculation, which is not ideal. I am not sure we can 100% do it in streaming mode as some parts of pretokenization might take at the outputs of previous modifications (Again not sure, but there's definitely some code looking backwards in the string to decide what to do.)

I have to mention https://github.com/microsoft/BlingFire that claims to be faster too. I have not used it since it got out nor checked accuracy. I don't think they output incoming offsets which can definitely help with speed. 

And last but not least, when checking for compliance, we like to use the XNLI dataset, as it contains tons of unicode, and has helped us catch many odd bugs in the past.