@benwtrent 2022-04-07
@Narsil , moved to a pre-tokenizer and did benches side by side for the optimized pretokenizer vs. not. There is almost no difference statistically.

Against "big.txt"

| Kind        | encode    | encode batch | encode, no cache | encode batch, no cache |
|-------------|-----------|--------------|------------------|------------------------|
| Regular     | 33.489 us | 9.8179 ms    | 42.064 us        | 12.178 ms              |
| "Optimized" | 31.539 us | 9.4979 ms    | 38.164 us        | 11.877 ms              |

Against "long-big.txt" (every 1k lines concatenated)

| Kind        | encode    | encode batch | encode, no cache | encode batch, no cache |
|-------------|-----------|--------------|------------------|------------------------|
| Regular     | 53.279 ms | 955.08 ms    | 66.155 ms        | 1.0737 s               |
| "Optimized" | 51.058 ms | 978.08 ms    | 61.584 ms        | 1.0252 s               |

Not 100% sure why I saw an improvement when I replaced the regex inline. I am guessing it was an artifact of running it on my machine. I double checked the inline numbers and there is no discernible difference now. 

One thing that has changed before my previous benches (showing the 10% speed improvement) and now is that there was a bug 🤦 . The bug was only adjusting the offset by `1` instead of `char.len()`. 

All in all, it shows my optimized code is not that optimized at all. `onig` is indeed about as fast as you can get it :D.