@benwtrent 2022-04-07
@Narsil as for BlingFire being faster...

Here is their bench:

https://github.com/microsoft/BlingFire/wiki/Comparing-performance-of-Bling-Fire-and-Hugging-Face-Tokenizers

From what I can tell, they are not using the "Fast" tokenizers? Of course their optimized code would be faster than python :/